+++ app/src/main/java/nju/com/piece/database/TagType.java
@@ -1,6 +1,7 @@
 package nju.com.piece.database;
 
-import nju.com.piece.entity.Timeline;
+
+import nju.com.piece.activity.TimeLineActivity;
 
 /**
  * Created by shen on 15/6/6.
@@ -8,27 +9,14 @@ import nju.com.piece.entity.Timeline;
 public enum TagType {
     work,relax;
 
-    public String getTimeField(){
-        String field = "";
-        switch (this) {
-            case work:
-                field = Timeline.WORK_TIME;
-                break;
-            case relax:
-                field = Timeline.RELAX_TIME;
-                break;
-        }
-        return field;
-    }
-
     public String getAllTimeField(){
         String field = "";
         switch (this) {
             case work:
-                field = Timeline.ALL_WORK_TIME;
+                field = TimeLineActivity.ALL_WORK_TIME;
                 break;
             case relax:
-                field = Timeline.ALL_RELAX_TIME;
+                field = TimeLineActivity.ALL_RELAX_TIME;
                 break;
         }
         return field;
